Make codec a Connection type parameter, split audit implementation
1. get rid of "workaround-audit-bug" "feature"

2. no longer use tokio_util::codec::{Decoder, Encoder}

tokio_util::codec is "designed" for bytestreams (and building "frames"
of messages on top), but we need to deal with datagrams (which still
can contain multiples messages, just not one message across multiple
datagrams).

This make it a little less "combinable", but we actually don't want
people to reuse these codecs on bytestream (and writing an adapter
wouldn't be that hard anyway).

Also we can use a fixed error type, making dealing with it a little
bit easier.